test training run

+++ init.sh
@@ -10,20 +10,25 @@ echo "Starting GPT project initialization..."
 echo "Installing dependencies from requirements.txt..."
 pip install -r requirements.txt &> /dev/null
 
-# Download data if folder doesn't exist
-# 1. check if the folder edu_fineweb10 exists
 if [ ! -d "edu_fineweb10" ]; then
     echo "Downloading data..."
-    python download_data.py &> /dev/null
-    # untar the file
-    tar -xzf fineweb.tar.gz
-    # clean up after
-    if [ -d "edu_fineweb10" ]; then
-        rm fineweb.tar.gz
+    python download_data.py
+
+    if [ -f "fineweb.tar.gz" ]; then
+        tar -xzf fineweb.tar.gz
+        if [ -d "edu_fineweb10" ]; then
+            rm fineweb.tar.gz
+        else
+            echo "Extraction failed. Directory not found after untar."
+            exit 1
+        fi
     else
-        echo "Data download failed. Please check the download_data.py script."
+        echo "Download failed. fineweb.tar.gz not found."
         exit 1
     fi
+else
+    echo "Data folder edu_fineweb10 already exists. Skipping download."
 fi
+
 echo "All tasks completed successfully!"
 echo "You can now train your model"
